requirements.txt
numpy==1.23.5
matplotlib==3.7.1
seaborn==0.12.2
jupyter==1.0.0
catboost==1.1.1
category_encoders==2.6.0
hyperopt==0.2.7
imbalanced_learn==0.10.1
imblearn==0.0
pandas==1.5.3
PyYAML==6.0
scikit_learn==1.2.2
xgboost==1.7.5
